requirements.txt
PyQt5>=5.15.0
pdfplumber>=0.7.0
Pillow>=9.0.0
chardet>=4.0.0 